refactor(p1): extract SmsAccessTools.kt — sms_list/read/send factories
Same factory shape as Contacts. Three internal createSmsXxxTool(context, deps)
top-level functions plus file-private querySms(context, input, previewOnly).

SystemAccessTools coordinator: 3 tool fields collapse to lazy delegators.

Side cleanup: drop now-unused imports `android.provider.Telephony` and
`android.telephony.SmsManager` from SystemAccessTools.kt.

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
